Allow HiveQueryRunner to persist data to a custom path
Add an optional argument to HiveQueryRunner to provide a path to read
and write the data and metadata to. The path must point to a writable
directory. On first run, HiveQueryRunner will create tpch and
tpch_bucketed tables and store them in the specified directory. On
subsequent runs using the same path, HiveQueryRunner will re-use the
data created before.

If no argument is provided, HiveQueryRunner will create a temporary
directory to store data and delete it before exiting.

+++ presto-main/src/main/java/com/facebook/presto/server/testing/TestingPrestoServer.java
@@ -79,8 +79,8 @@ import javax.annotation.concurrent.GuardedBy;
 
 import java.io.Closeable;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.net.URI;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -98,6 +98,7 @@ import static com.google.common.io.MoreFiles.deleteRecursively;
 import static com.google.common.io.RecursiveDeleteOption.ALLOW_INSECURE;
 import static io.airlift.discovery.client.ServiceAnnouncement.serviceAnnouncement;
 import static java.lang.Integer.parseInt;
+import static java.nio.file.Files.createTempDirectory;
 import static java.nio.file.Files.isDirectory;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
@@ -107,6 +108,7 @@ public class TestingPrestoServer
 {
     private final Injector injector;
     private final Path baseDataDir;
+    private final boolean preserveData;
     private final LifeCycleManager lifeCycleManager;
     private final PluginManager pluginManager;
     private final ConnectorManager connectorManager;
@@ -171,7 +173,8 @@ public class TestingPrestoServer
         this(true, ImmutableMap.of(), null, null, new SqlParserOptions(), additionalModules);
     }
 
-    public TestingPrestoServer(boolean coordinator,
+    public TestingPrestoServer(
+            boolean coordinator,
             Map<String, String> properties,
             String environment,
             URI discoveryUri,
@@ -179,8 +182,23 @@ public class TestingPrestoServer
             List<Module> additionalModules)
             throws Exception
     {
+        this(coordinator, properties, environment, discoveryUri, parserOptions, additionalModules, Optional.empty());
+    }
+
+    public TestingPrestoServer(
+            boolean coordinator,
+            Map<String, String> properties,
+            String environment,
+            URI discoveryUri,
+            SqlParserOptions parserOptions,
+            List<Module> additionalModules,
+            Optional<Path> baseDataDir)
+            throws Exception
+    {
         this.coordinator = coordinator;
-        baseDataDir = Files.createTempDirectory("PrestoTest");
+
+        this.baseDataDir = baseDataDir.orElseGet(TestingPrestoServer::tempDirectory);
+        this.preserveData = baseDataDir.isPresent();
 
         properties = new HashMap<>(properties);
         String coordinatorPort = properties.remove("http-server.http.port");
@@ -312,7 +330,7 @@ public class TestingPrestoServer
             throw new RuntimeException(e);
         }
         finally {
-            if (isDirectory(baseDataDir)) {
+            if (isDirectory(baseDataDir) && !preserveData) {
                 deleteRecursively(baseDataDir, ALLOW_INSECURE);
             }
         }
@@ -507,5 +525,15 @@ public class TestingPrestoServer
             }
         }
         throw new RuntimeException("Presto announcement not found: " + announcements);
+    }
+
+    private static Path tempDirectory()
+    {
+        try {
+            return createTempDirectory("PrestoTest");
+        }
+        catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
     }
 }
